# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Reservations - API",
        "version": "1.0"
    },
    "paths": {
...
        "/api/movie/{movieId}/reservations": {
            "get": {
                "tags": [
                    "Reservations"
                ],
                "summary": "Gets the list of reservations by movie",
                "description": "Returns the list of reservations by movie with optional pagination",
                "operationId": "getReservationsListByMovie",
                "parameters": [
                    {
                        "name": "perPage",
                        "in": "query",
                        "description": "Number of cinemas per page",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 10
                        }
                    },
                    {
                        "name": "movieId",
                        "in": "path",
                        "description": "ID of movie to return reservations",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "No reservations found"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            },
            "post": {
                "tags": [
                    "Reservations"
                ],
                "summary": "Store a new reservation by movie",
                "description": "Stores a new reservation by movie and returns the reservation data",
                "operationId": "storeReservationByMovie",
                "parameters": [
                    {
                        "name": "movieId",
                        "in": "path",
                        "description": "ID of cinema to store room",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Reservation data",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "seance_id",
                                    "seat"
                                ],
                                "properties": {
                                    "seance_id": {
                                        "type": "string",
                                        "format": "uuid",
                                        "example": "3434-4343-4343-4343"
                                    },
                                    "seat": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2021-09-15"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "string",
                                            "format": "uuid",
                                            "example": "3434-4343-4343-4343"
                                        },
                                        "seance_id": {
                                            "type": "string",
                                            "format": "uuid",
                                            "example": "3434-4343-4343-4343"
                                        },
                                        "seat": {
                                            "type": "integer",
                                            "format": "int",
                                            "example": "34"
                                        },
                                        "rank": {
                                            "type": "integer",
                                            "format": "int",
                                            "example": "3"
                                        },
                                        "status": {
                                            "type": "string",
                                            "format": "string",
                                            "example": "open"
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "format": "date",
                                            "example": "2021-09-15"
                                        },
                                        "updated_at": {
                                            "type": "string",
                                            "format": "date",
                                            "example": "2021-09-15"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Invalid input data"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            }
        },
...
        "/api/reservations/{reservationId}/confirm": {
            "put": {
                "tags": [
                    "Reservations"
                ],
                "summary": "Confirm an existing reservation",
                "description": "Confirms and returns a reservation data",
                "operationId": "confirmReservation",
                "parameters": [
                    {
                        "name": "reservationId",
                        "in": "path",
                        "description": "ID of reservation that needs to be confirmed",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Reservation confirmed",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "string",
                                            "format": "uuid",
                                            "example": "3434-4343-4343-4343"
                                        },
                                        "seance_id": {
                                            "type": "string",
                                            "format": "uuid",
                                            "example": "3434-4343-4343-4343"
                                        },
                                        "seat": {
                                            "type": "integer",
                                            "format": "int",
                                            "example": "34"
                                        },
                                        "rank": {
                                            "type": "integer",
                                            "format": "int",
                                            "example": "3"
                                        },
                                        "status": {
                                            "type": "string",
                                            "format": "string",
                                            "example": "open"
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "format": "date",
                                            "example": "2021-09-15"
                                        },
                                        "updated_at": {
                                            "type": "string",
                                            "format": "date",
                                            "example": "2021-09-15"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid ID supplied"
                    },
                    "404": {
                        "description": "Reservation not found"
                    },
                    "405": {
                        "description": "Validation exception"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            }
        },
...
    }
}